requirements.txt
asn1crypto==0.24.0
base58==0.2.5
bitcash==0.5.2
bottle==0.12.13
cashaddress==1.0.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coincurve==7.1.0
coverage==4.5.1
Django==2.0.5
django-appconf==1.0.2
django-autoslug==1.9.3
django-compressor==2.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-haystack==2.8.1
django-ipware==2.1.0
django-pagedown==1.0.4
django-uuslug==1.1.8
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.6
jieba==0.39
Markdown==2.6.11
Pillow==5.1.0
pycparser==2.18
PyMySQL==0.8.0
python-slugify==1.2.5
pytz==2018.4
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.22
urllib3==1.22
WeRoBot==1.4.0
xmltodict==0.11.0
